solving #19 issue. Fixed the error created upon the release of the emojis package version.

+++ py_heideltime/py_heideltime.py
@@ -58,11 +58,12 @@ def py_heideltime(text, language='English', date_granularity='full', document_ty
         new_text = ''.join(new_text_list)
         tagged_text = ''.join(tagged_text_list)
         ExecTimeDictionary={'heideltime_processing': heideltime_processing_time-sum(py_heideltime_text_normalization), 'py_heideltime_text_normalization': sum(py_heideltime_text_normalization)}
-        return [dates_results, new_text, tagged_text, ExecTimeDictionary]
-    finally:
-        shutil.rmtree(directory_name) #remove folder and files that were processed by heideltime
+        if os.path.exists(directory_name):
+            shutil.rmtree(directory_name) #remove folder and files that were processed by heideltime
         os.remove('config.props')   #remove config.props files
-
+        return [dates_results, new_text, tagged_text, ExecTimeDictionary]
+    except Exception as e:
+        print("Error: " + str(e))
 
 def create_txt_files(text, directory_name):
     chunkSize = 30000 #30000 chars
@@ -187,11 +188,9 @@ def remove_emoji(text):
     return emoji.get_emoji_regexp().sub(u'', text)
 
 def text_has_emoji(text):
-    for character in text:
-        if character in emoji.UNICODE_EMOJI:
+    if  emoji.distinct_emoji_list(text):
             return True
     return False
-
 
 def pre_process_text(text):
     if text_has_emoji(text):
